Refactor : 관련 있는 dto 내부 record로 선언
