+++ src/main/java/org/guzzing/studayserver/domain/calendar/controller/dto/request/AcademyCalendarCreateRequest.java
@@ -5,8 +5,11 @@ import jakarta.validation.constraints.*;
 import org.guzzing.studayserver.domain.calendar.controller.dto.request.validation.ValidEnum;
 import org.guzzing.studayserver.domain.calendar.model.Periodicity;
 import org.guzzing.studayserver.domain.calendar.service.dto.param.AcademyCalendarCreateParam;
+import org.guzzing.studayserver.domain.calendar.service.dto.param.LessonScheduleParam;
 
+import java.time.DayOfWeek;
 import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.List;
 
 public record AcademyCalendarCreateRequest(
@@ -49,6 +52,23 @@ public record AcademyCalendarCreateRequest(
                 request.memo,
                 Periodicity.valueOf(request.periodicity)
         );
+    }
+
+    public record LessonScheduleCreateRequest(
+            @NotBlank
+            @ValidEnum(enumClass = DayOfWeek.class, message = "올바른 요일 표기가 아닙니다.")
+            String dayOfWeek,
+
+            @Valid
+            LessonTime lessonTime
+    ) {
+        public static LessonScheduleParam to(LessonScheduleCreateRequest request) {
+            return new LessonScheduleParam(
+                    DayOfWeek.valueOf(request.dayOfWeek()),
+                    LocalTime.parse(request.lessonTime().getLessonStartTime()),
+                    LocalTime.parse(request.lessonTime().getLessonEndTime())
+            );
+        }
     }
 
 }
